+++ HISTORY.md
@@ -11,6 +11,7 @@
   bwa with multiple cores.
 - Tumor-only prioritization uses vcfanno output instead of GEMINI,
   allowing use without needing to build a full GEMINI database.
+- Use samtools multicore indexing, replacing sambamba multicore index.
 - Multicore base quality score recalibration with GATK4 and Sentieon.
 - GATK4: add support for gVCF based joint calling.
 - GATK4: fix option usage for gVCF creation with HaplotypeCaller